emotonalttsapp/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
edge-tts
transformers
torch
pydub
python-multipart
tf-keras
numpy<2.0